Fix para el bug de Cantidad para valores decimales en Venta Mayorista y Compra

+++ NaturalFrut/App_BLL/VentaMayoristaLogic.cs
@@ -140,7 +140,7 @@ namespace NaturalFrut.App_BLL
             return listaRP.GetAll().ToList();
         }
 
-        public ListaPrecio CalcularImporteSegunCliente(int clienteID, int productoID, int cantidad)
+        public ListaPrecio CalcularImporteSegunCliente(int clienteID, int productoID, double cantidad)
         {
 
             var cliente = clienteRP.GetByID(clienteID);
